onedrive: default onedrive personal to QuickXorHash and implement --onedrive-hash-type
Before this change the hash used for Onedrive Personal was SHA1.
However over the next few months Microsoft are phasing out SHA1 hashes
in favour of QuickXorHash in Onedrive Personal. Onedrive Business and
Sharepoint remain using QuickXorHash as before.

This also implement the --onedrive-hash-type flag (and config option)
so that the hash can be selected in the transition period.